+++ packages/workflow/src/NodeHelpers.ts
@@ -6,6 +6,7 @@
 import get from 'lodash/get';
 import isEqual from 'lodash/isEqual';
 
+import { EXECUTE_WORKFLOW_NODE_TYPE } from './Constants';
 import { ApplicationError } from './errors/application.error';
 import { NodeConnectionTypes } from './Interfaces';
 import type {
@@ -39,6 +40,7 @@ import type {
 import { validateFilterParameter } from './NodeParameters/FilterParameter';
 import {
 	isFilterValue,
+	isResourceLocatorValue,
 	isResourceMapperValue,
 	isValidResourceLocatorParameterValue,
 } from './type-guards';
@@ -1561,4 +1563,18 @@ export function isExecutable(workflow: Workflow, node: INode, nodeTypeData: INod
 		isTriggerNode(nodeTypeData) ||
 		nodeTypeData.usableAsTool === true
 	);
+}
+
+/**
+ * Attempts to retrieve the ID of a subworkflow from a execute workflow node.
+ */
+export function getSubworkflowId(node: INode): string | undefined {
+	if (
+		node &&
+		node.type === EXECUTE_WORKFLOW_NODE_TYPE &&
+		isResourceLocatorValue(node.parameters.workflowId)
+	) {
+		return node.parameters.workflowId.value as string;
+	}
+	return;
 }
